refactor: 지도 ViewModel 테스트를 관심사별로 분리
Refs #205

- CollectionSpotMapViewModelTest 단일 파일을 관심사별 테스트 파일로 분리
- 현재 위치, 초기 진입, 검색, 캐시, 선택 상태, 즐겨찾기 테스트 범위 정리
- 공통 ViewModel 생성 helper와 fake repository/provider/checker를 fixture로 분리
- 기존 테스트명과 검증 의미 유지
- 프로덕션 코드 변경 없이 test source set만 정리

검증:
- ./gradlew.bat :presentation:testDebugUnitTest
- ./gradlew.bat :presentation:compileDebugKotlin
- ./gradlew.bat :app:assembleDebug
- git diff --check
